+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 #1.2.4
 - Added separate configs for icons and labels opacity
-- Added 'Circle location' config (always under cursor / always corner / Under cursor + in corner when no space)
+- Added 'Circle location' config (always under cursor / always corner / under cursor + in corner when no space / under cursor + shift when no space)
 - Changed some default configs
 - Small options page improvements